deploy1.3.1_Transcriptions2JSON_saveDB/backend/apiTrans2JsonDicts_forDeloy1.3/deploy1.1_H3_5PromptDetAndScoring/requirements.txt
pyyaml 
python-dotenv